feat: Unit-based stock management for bulk products
- Changed stockQuantity from double to int (complete unopened units)
- Added currentUnitRemaining (double) to track opened unit
- Portions now deducted from opened unit, opening new units as needed
- Updated stock verification in shop and cart controllers
- Stock alerts now work on unit count instead of portions

+++ lib/app/data/repositories/product_repository.dart
@@ -87,8 +87,8 @@ class ProductRepository {
     required String description,
     required double price,
     required int categoryId,
-    required double stockQuantity,
-    required double minStockAlert,
+    required int stockQuantity,
+    required int minStockAlert,
     String? imageUrl,
     bool isBulkProduct = false,
     String? bulkUnit,
@@ -120,12 +120,12 @@ class ProductRepository {
     required String description,
     required double price,
     required int categoryId,
-    required double minStockAlert,
+    required int minStockAlert,
     String? imageUrl,
     bool? isBulkProduct,
     String? bulkUnit,
     double? bulkTotalQuantity,
-    double? stockQuantity,
+    int? stockQuantity,
   }) async {
     try {
       return await _client.product.updateProduct(
@@ -179,7 +179,7 @@ class ProductRepository {
   }
 
   /// Update product stock quantity (admin only)
-  Future<dynamic> updateStock(int productId, double newStockQuantity) async {
+  Future<dynamic> updateStock(int productId, int newStockQuantity) async {
     try {
       final result = await _client.product.updateStock(
         productId,
